Finished the Game.
Added the checkspelling function to make sure that the users Input is correct.
Added checkWinner function which checks if either the user of the pc won the round.
Added playGame function which repeatidly plays the game until User or Pc has gotten a score of 3.
Added a announceWinner function that announces the Winner once either the PC or User has gotten a total score of 3